+++ .gitignore
@@ -15,3 +15,6 @@ bin/
 
 /text-ui-test/ACTUAL.TXT
 text-ui-test/EXPECTED-UNIX.TXT
+text-ui-test/EXPECTED.TXT
+
+data/Bimo.txt
