+++ build.gradle
@@ -18,7 +18,6 @@ test {
 
     testLogging {
         events "passed", "skipped", "failed"
-
         showExceptions true
         exceptionFormat "full"
         showCauses true
@@ -28,7 +27,7 @@ test {
 }
 
 application {
-    mainClass.set("Bimo.Bimo")
+    mainClass.set("bimo.Bimo")
 }
 
 shadowJar {
