+++ src/main/java/bimo/BimoException.java
@@ -1,4 +1,4 @@
-package Bimo;
+package bimo;
 
 public class BimoException extends Exception {
     public BimoException(String message) {
