+++ src/main/java/bimo/Ui.java
@@ -1,8 +1,5 @@
-package Bimo;
+package bimo;
 
-import java.time.LocalDate;
-import java.time.format.DateTimeParseException;
-import java.util.ArrayList;
 import java.util.Scanner;
 
 public class Ui {
