+++ src/main/java/bimo/command/Command.java
@@ -1,8 +1,8 @@
-package Bimo.Command;
+package bimo.command;
 
-import Bimo.Storage;
-import Bimo.TaskList;
-import Bimo.Ui;
+import bimo.Storage;
+import bimo.TaskList;
+import bimo.Ui;
 
 public abstract class Command {
     private boolean isQuit = false;
