+++ src/main/java/bimo/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package Bimo.Tasks;
+package bimo.tasks;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
