+++ src/main/java/bimo/tasks/Event.java
@@ -1,4 +1,4 @@
-package Bimo.Tasks;
+package bimo.tasks;
 
 
 import java.time.LocalDate;
