+++ src/main/java/bimo/tasks/ToDo.java
@@ -1,4 +1,4 @@
-package Bimo.Tasks;
+package bimo.tasks;
 
 public class ToDo extends Task {
