+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
     ___________________________________
-    Hello! I'm Bimo
+    Hello! I'm bimo
     What can I do for you?
     ___________________________________
     ___________________________________
